requirements.txt
torch==1.12.0
torchvision==0.13.0
pandas==2.0.3
numpy==1.24.4
Pillow==9.5.0
scikit-learn==1.3.0
matplotlib
ftfy 
regex 
tqdm